requirements.txt
streamlit
pandas
sqlalchemy
python-dotenv
psycopg2-binary
faker
requests 